Refactor order-related functionality and add filtering capabilities
- Updated  to select only 'id' and 'name' fields for leaf categories.
- Changed validation rules in , , , , and  to use 'numeric' instead of 'integer' for IDs.
- Added new controllers: , , and  to handle paginated order retrieval and filter dialog data.
- Implemented request validation in  to support filtering by various criteria.
- Created resource classes  and  for structured order responses.
- Added tests for paginated orders retrieval and filtering by category.

+++ app/Http/Requests/Categories/UpdateParentCategoryRequest.php
@@ -33,7 +33,7 @@ class UpdateParentCategoryRequest extends FormRequest
         return [
             "new_parent_id" => [
                 'nullable',
-                'integer',
+                'numeric',
                 Rule::exists('categories', 'id')
             ],
         ];
